requirement.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
openai
python-dotenv
python-telegram-bot
requests